requirements.txt
google-generativeai
python-telegram-bot==20.0
python-dotenv